requirements.txt
undetected-chromedriver>=3.5.0
selenium>=4.15.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
